[cern] ci: Keep images from being deleted periodically
We have a script that deletes images every 10 mins and
it deletes the "intermediate" kubernetes-master and node
which are used for the rest of the kube components.

Start a container which sleeps for 3600 secs to prevent
image deletion.

Also, remove all rmi commands since the clean up job
will do it.